requirements.txt
wradlib
requests==2.24.0
numpy
rfc3339
schedule
python-confluent-kafka
